SN-612 Adding distribution files and a Makefile to help with publishing versions
