+++ .gitignore
@@ -272,3 +272,5 @@ $RECYCLE.BIN/
 *.lnk
 
 # End of https://www.gitignore.io/api/linux,macos,python,windows,intellij+iml
+
+venv/**
