+++ requirements.txt
@@ -1,0 +1,1 @@
+requests~=2.22.0
