--- securenative/errors.py
@@ -1,2 +1,0 @@
-class MissingApiKeyError(ValueError):
-    pass
