+++ securenative/http_client_test.py
@@ -1,7 +1,7 @@
 import json
 import unittest
 
-from securenative.config import sdk_version
+from config import sdk_version
 from securenative.http_client import HttpClient
